refactor: split lib utilities into modules
